% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sim.R
\name{cluster_sim}
\alias{cluster_sim}
\title{Calculate expected coverage of several alternative intervals for
 marginal coefficients in linear regression with clustered coefficients.}
\usage{
cluster_sim(beta, Sigma, err.sd = 1, n.samp = 100, n.rep = 100,
  seed = NULL, parallel = FALSE)
}
\arguments{
\item{beta}{List of regression coefficients.
This should be a least with K elements where K is the number of blocks}

\item{Sigma}{List of correaltion matrices, one for each block}

\item{which.sample}{Indices of individuals who are sampled}
}
\value{
A list with elements:
\describe{
\item{\code{COVERAGE}}{A 5 by p by n array giving the coverage at each simulation.
\code{COVERAGE[i, j, k]} is either 0 or 1 indicating weather the interval formed using
method i for rank j in simulation k covered its respective parameter}
\item{\code{WIDTH}}{ A 5 by p by n array giving the width of each interval in each simulation.}
 \item{\code{simnames}}{Name of each type of the 5 types of confidence intervals}
}
}
\description{
This function can be used to generate simulations in section 3.1
}
