% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sim.R
\name{example_sim}
\alias{example_sim}
\title{Calculate expected coverage of several alternative intervals given a set of parameter values.}
\usage{
example_sim(theta, n = 200, use.abs = TRUE)
}
\arguments{
\item{theta}{Vector of parameter values (length p)}

\item{n}{Number of simulations to run.}

\item{use.abs}{Rank estimates by absolute value for bootstrapping methods.}
}
\value{
A list with elements:
\describe{
\item{\code{COVERAGE}}{A 5 by p by n array giving the coverage at each simulation.
\code{COVERAGE[i, j, k]} is either 0 or 1 indicating weather the interval formed using
method i for rank j in simulation k covered its respective parameter}
\item{\code{WIDTH}}{ A 5 by p by n array giving the width of each interval in each simulation.}
 \item{\code{Z}}{A p by n matrix of all of the simulated data}
 \item{\code{simnames}}{Name of each type of the 5 types of confidence intervals}
}
}
\description{
This function can be used to generate the example shown in figure 1.
}
