% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sims.R
\name{plot_cis}
\alias{plot_cis}
\title{Plot a set of confidence intervals and the true parameters}
\usage{
plot_cis(rank, ci, truth, prop = 1, plot.truth = FALSE)
}
\arguments{
\item{rank}{Rank of each parameter (will by x-axis)}

\item{ci}{p by 2 matrix or data frame giving the confidence interval for each parameter}

\item{truth}{True parameters (vector of length p)}

\item{prop}{Proportion of praameters to plot. Should be less than 1.}

\item{plot.truth}{Logical, should the trute parameters be shown on the plot}
}
\value{
A ggplot object
}
\description{
Plot a set of confidence intervals and the true parameters
}
